@@ -4,22 +4,23 @@
 	"description": "A tool to edit videos",
 	"private": true,
 	"engines": {
-		"node": "7.10.1",
+		"node": ">=14.9.3",
 		"npm": "6.2.0"
 	},
 	"dependencies": {
 		"@wikimedia/react.i18n": "^2.0.3",
-		"axios": "^0.21.4",
-		"bootstrap": "^4.6.0",
-		"concurrently": "^6.3.0",
+		"axios": "^0.27.2",
+		"bootstrap": "^5.1.3",
+		"concurrently": "^7.2.1",
+		"cross-env": "^7.0.3",
 		"popup-tools": "^1.0.2",
-		"react": "^17.0.2",
-		"react-bootstrap": "^1.6.4",
-		"react-bootstrap-icons": "^1.5.0",
-		"react-dom": "^17.0.2",
-		"react-router-dom": "^5.3.0",
-		"react-scripts": "^4.0.3",
-		"socket.io-client": "^4.3.1"
+		"react": "^18.1.0",
+		"react-bootstrap": "^2.4.0",
+		"react-bootstrap-icons": "^1.8.2",
+		"react-dom": "^18.1.0",
+		"react-router-dom": "^6.3.0",
+		"react-scripts": "^5.0.1",
+		"socket.io-client": "^4.5.1"
 	},
 	"keywords": [
 		"VideoCutTool",
@@ -35,7 +36,7 @@
 	"bugs": {
 		"url": "https://phabricator.wikimedia.org/tag/videocuttool/"
 	},
-	"homepage": "https://videocuttool.wmflabs.org",
+	"homepage": "https://videocuttool.wmcloud.org",
 	"scripts": {
 		"start": "react-scripts start",
 		"build": "react-scripts build",
@@ -43,7 +44,7 @@
 		"eject": "react-scripts eject",
 		"postinstall": "npm --prefix ./server install",
 		"serve": "npm run serve --prefix server",
-		"dev": "concurrently \"npm run start\" \"npm run serve\"",
+		"dev": "cross-env NODE_ENV=development concurrently \"npm run start\" \"npm run serve\"",
 		"prd": "npm run start:prod --prefix server",
 		"server": "node ./server/index.js"
 	},
@@ -54,15 +55,18 @@
 		"not op_mini all"
 	],
 	"devDependencies": {
-		"eslint": "^7.32.0",
-		"eslint-config-airbnb": "^18.2.1",
-		"eslint-import-resolver-alias": "^1.1.2",
-		"eslint-plugin-import": "^2.24.2",
-		"eslint-plugin-jsx-a11y": "^6.4.1",
-		"eslint-plugin-react": "^7.26.1",
-		"eslint-plugin-react-hooks": "^4.2.0",
-		"stylelint": "^13.13.1",
-		"stylelint-config-standard": "^22.0.0",
-		"stylelint-scss": "^3.21.0"
+		"eslint": "8.16.0",
+		"eslint-config-airbnb": "19.0.4",
+		"eslint-import-resolver-alias": "1.1.2",
+		"eslint-plugin-import": "2.26.0",
+		"eslint-plugin-jsx-a11y": "6.5.1",
+		"eslint-plugin-react": "7.30.0",
+		"eslint-plugin-react-hooks": "4.5.0",
+		"postcss-flexbugs-fixes": "^5.0.2",
+		"postcss-normalize": "^10.0.1",
+		"postcss-preset-env": "^7.6.0",
+		"stylelint": "14.8.5",
+		"stylelint-config-standard": "25.0.0",
+		"stylelint-scss": "4.2.0"
 	}
 }
