@@ -23,15 +23,13 @@ function connectMongoDB(retry = 0) {
 	};
 
 	mongoose
-
 		.connect(config.DB_CONNECTION_URL, option)
 		.then(() => console.log('MongoDB Connected'))
 		.catch(err => {
 			console.log('error', '--------------------');
 			console.log(err.message);
 			console.log(config.DB_CONNECTION_URL);
-			console.log('>>>>>>>>>>>>>>>>>>>>', `Reconnecting to MongoDB ${retry}`);
-			connectMongoDB(retry + 1);
+			console.log(`Reconnecting to MongoDB ${retry}`);
 		});
 }
 
@@ -39,12 +37,22 @@ connectMongoDB();
 
 const app = express();
 
-const __dirname = path.resolve();
+const __dirname =
+	process.env.NODE_ENV === 'production' ? `${path.resolve()}/` : `${path.resolve()}/server/`;
 
-app.use('/public', express.static(path.join(__dirname, 'public')));
+app.use('/api/public', express.static(path.join(__dirname, 'public')));
 
 app.set('view engine', 'pug');
 app.set('views', path.join(__dirname, 'views'));
+
+app.use(
+	fileUpload({
+		useTempFiles: true,
+		tempFileDir: 'tmp/', // so that they're publicly accessible
+		limits: { fileSize: 500 * 1024 * 1024 },
+		abortOnLimit: true
+	})
+);
 
 app.use(logger('dev'));
 app.use(express.json({ limit: '500mb' }));
@@ -53,15 +61,6 @@ app.use(cookieParser());
 
 // Use CORS and File Upload modules here
 app.use(cors());
-
-app.use(
-	fileUpload({
-		useTempFiles: true,
-		tempFileDir: 'tmp/', // so that they're publicly accessable
-		limits: { fileSize: 500 * 1024 * 1024 },
-		abortOnLimit: true
-	})
-);
 
 app.use(
 	session({
@@ -76,11 +75,19 @@ app.get('/', (req, res) => {
 	res.json({ data: 'Homepage' });
 });
 
-app.get('/error', (req, res) => {
+app.get('/api/', (req, res) => {
+	res.json({ data: 'Back-end is up' });
+});
+
+app.get('/api/error', (req, res) => {
 	res.render('error', { error_message: req.session.error_message });
 });
 
-app.get('/login', (req, res) => {
+app.get('/test-auth', (req, res) => {
+	+res.sendFile(path.join(`${__dirname}/test-auth.html`));
+});
+
+app.get('/api/login', (req, res) => {
 	const baseUrl = 'https://commons.wikimedia.org';
 	const endpoint = '/w/rest.php/oauth2/authorize';
 
@@ -91,7 +98,7 @@ app.get('/login', (req, res) => {
 	res.send(res.redirect(url));
 });
 
-app.get('/auth/mediawiki/callback', auth, async (req, res) => {
+app.get('/api/auth/mediawiki/callback', auth, async (req, res) => {
 	const {
 		refresh_token: refreshToken,
 		profile,
@@ -112,24 +119,26 @@ app.get('/auth/mediawiki/callback', auth, async (req, res) => {
 		};
 		res.end(PopupTools.popupResponse({ user: returnUserDocData }));
 	} catch (err) {
+		console.log('************');
+		console.log(err);
 		const error = err.toJSON();
 		req.session.error_message = error.message;
 		res.redirect('/error');
 	}
 });
 
-app.get('/logout', (req, res) => {
+app.get('/api/logout', (req, res) => {
 	delete req.session.user;
 	res.redirect('/');
 });
 
-app.post('/process', processVideo);
-app.post('/upload', uploadVideos);
-app.get('/download/:videopath', downloadVideo);
+app.post('/api/process', processVideo);
+app.post('/api/upload', uploadVideos);
+app.get('/api/download/:videopath', downloadVideo);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
-	const err = new Error('Not Found');
+	const err = new Error(`Not Found${req.originalUrl}`);
 	err.status = 404;
 	next(err);
 });
