@@ -14,7 +14,10 @@ server.listen(PORT);
 
 const io = new Server(server, {
 	cors: {
-		origin: 'http://localhost:3000',
+		origin:
+			process.env.NODE_ENV !== 'development'
+				? 'https://videocuttool.wmcloud.org'
+				: 'http://localhost:3000',
 		methods: ['GET', 'POST'],
 		credentials: true
 	},
@@ -23,6 +26,7 @@ const io = new Server(server, {
 });
 
 io.sockets.on('connection', socket => {
+	console.log('CONNECTED', socket.id);
 	app.set('socketid', socket.id);
 	socketController(socket, io);
 });
