@@ -1,15 +1,15 @@
-import React, { useState, useRef, useContext } from 'react';
+import React, { useState, useRef, useContext, useEffect } from 'react';
 import { Message } from '@wikimedia/react.i18n';
 import { Form, FormLabel } from 'react-bootstrap';
-import path from 'path';
 import axios from 'axios';
 import { AppContext } from '../context';
 
-function UrlBox() {
+function UrlBox(props) {
 	const { updateAppState } = useContext(AppContext);
 
-	const allowedExtensions = '.mp4,.webm,.mov,.flv,.ogv';
+	const allowedExtensions = 'mp4,webm,mov,flv,ogv';
 	const [mouseHover, setMouseHover] = useState(false);
+	const [title, setTitle] = useState('');
 	const fileUpload = useRef(null);
 	const dragEnter = () => {
 		setMouseHover(true);
@@ -29,21 +29,65 @@ function UrlBox() {
 		onFileUpload(e);
 	};
 
+	useEffect(() => {
+		setTitle(props.title);
+		checkFileExist(props.title);
+	}, [props.title]);
+
 	const onFileUpload = e => {
 		const files = (e.dataTransfer && e.dataTransfer.files) || e.nativeEvent.target.files;
 		if (files.length === 0) {
 			return;
 		}
 
-		const fileExt = path.extname(files[0].name);
-
-		/*		if (allowedExtensions.split(',').indexOf(fileExt) === -1) {
+		const fileExt = files[0].name.split('.').pop();
+		if (allowedExtensions.split(',').indexOf(fileExt) === -1) {
+			alert('File extension not allowed');
 			return;
-		} */
+		}
+
+		updateAppState({
+			current_step: 2,
+			video_url: URL.createObjectURL(files[0]),
+			file: files[0],
+			video_details: {
+				title: files[0].name.replace(/\s/g, '_')
+			}
+		});
 	};
 
-	const onUrlInput = e => {
-		checkFileExist(e.target.value);
+	/**
+	 * Get video date from either commons site or user uploads
+	 *
+	 * @param {string} videoUrl Video Path
+	 * @returns
+	 */
+	const retrieveVideoData = videoUrl => {
+		const splitUrl = videoUrl.split('/');
+		if (!videoUrl.includes('commons.wikimedia.org')) {
+			return;
+		}
+		axios
+			.get(
+				`https://commons.wikimedia.org/w/api.php?action=query&format=json&prop=videoinfo&titles=${
+					splitUrl[splitUrl.length - 1]
+				}&viprop=user%7Curl%7Ccanonicaltitle%7Ccomment%7Curl&origin=*`
+			)
+			.then(response => {
+				const { pages } = response.data.query;
+				if (Object.keys(pages)[0] !== '-1') {
+					const { user, canonicaltitle, comment, url } = pages[Object.keys(pages)[0]].videoinfo[0];
+					updateAppState({
+						current_step: 2,
+						video_url: url,
+						video_details: {
+							author: user,
+							title: decodeURIComponent(canonicaltitle.slice(5)).replace(/\s/g, '_'),
+							comment
+						}
+					});
+				}
+			});
 	};
 
 	/**
@@ -78,46 +122,17 @@ function UrlBox() {
 					return;
 				}
 
-				// File exists, retrive video data
-				retriveVideoData(filePath);
+				// File exists, retrieve video data
+				retrieveVideoData(filePath);
 			})
 			.catch(error => {
 				console.log(error);
 			});
 	};
 
-	/**
-	 * Get video date from either commons site or user uploads
-	 *
-	 * @param {string} videoUrl Video Path
-	 * @returns
-	 */
-	const retriveVideoData = videoUrl => {
-		const splitUrl = videoUrl.split('/');
-		if (!videoUrl.includes('commons.wikimedia.org')) {
-			return;
-		}
-		axios
-			.get(
-				`https://commons.wikimedia.org/w/api.php?action=query&format=json&prop=videoinfo&titles=${
-					splitUrl[splitUrl.length - 1]
-				}&viprop=user%7Curl%7Ccanonicaltitle%7Ccomment%7Curl&origin=*`
-			)
-			.then(response => {
-				const { pages } = response.data.query;
-				if (Object.keys(pages)[0] !== '-1') {
-					const { user, canonicaltitle, comment, url } = pages[Object.keys(pages)[0]].videoinfo[0];
-					updateAppState({
-						current_step: 2,
-						video_url: url,
-						video_details: {
-							author: user,
-							title: decodeURIComponent(canonicaltitle.slice(5)).replace(/\s/g, '_'),
-							comment
-						}
-					});
-				}
-			});
+	const onUrlInput = e => {
+		setTitle(e.target.value);
+		checkFileExist(e.target.value);
 	};
 
 	return (
@@ -149,6 +164,7 @@ function UrlBox() {
 					placeholder="https://commons.wikimedia.org/wiki/File:video.webm"
 					onChange={onUrlInput}
 					autoComplete="true"
+					value={title}
 				/>
 			</div>
 		</div>
