@@ -3,4 +3,6 @@ import { io } from 'socket.io-client';
 const { backend_url, socket_io_url, socket_io_path } = require('../env')();
 
 export const API_URL = backend_url;
-export const socket = io(socket_io_url, { path: socket_io_path });
+export const socket = io(
+	process.env.NODE_ENV === 'development' ? 'http://localhost:4000' : undefined
+);
